Further modularize cytobands code
